Apply improved types & architecture to Filecoin
